+++ types/tiletype.d.ts
@@ -20,9 +20,9 @@ declare module "tiletype" {
      * * webp
      *
      * @param {Buffer} buffer input
-     * @returns {String|boolean} identifier
+     * @returns {String} identifier
      */
-    export function type(buffer: Buffer): extensions | boolean
+    export function type(buffer: Buffer): extensions
 
     /**
      * Return headers - Content-Type and Content-Encoding -
@@ -38,7 +38,7 @@ declare module "tiletype" {
      * returned as a [x, y] array.
      *
      * @param {Buffer} buffer input
-     * @returns {Array<number>|boolean} dimensions
+     * @returns {Array<number>} dimensions
      */
-    export function dimensions(buffer: Buffer): [number, number] | boolean
+    export function dimensions(buffer: Buffer): [number, number]
 }