+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 
 # Changelog
 
+## 1.0.2 - 2017-1-1
+
+- Handle metadata via class attributes
+- Detect image type using tiletype
+
 ## 1.0.0 - 2016-12-30
 
 - Initialize library